type casts

+++ datamodel/app/sql_functions/symbology_functions.sql
@@ -396,7 +396,7 @@ BEGIN
   SET _depth = calc_depth
   FROM (
     SELECT ws.obj_id,
-      nullif(MAX(co.level),0) - COALESCE(nullif(MIN(wn.bottom_level),0), nullif(MIN(RP.level),0)) as calc_depth
+      nullif(MAX(co.level),0)::numeric - COALESCE(nullif(MIN(wn.bottom_level),0), nullif(MIN(RP.level),0))::numeric as calc_depth
       FROM tww_od.wastewater_structure ws
       LEFT JOIN tww_od.cover co on ws.fk_main_cover = co.obj_id
       LEFT JOIN tww_od.wastewater_networkelement ne ON ne.fk_wastewater_structure = ws.obj_id
